+++ Changelog.md
@@ -2,6 +2,7 @@
 
 __New Features__
 - Updates to HPXML v4.2-rc3.
+- Allows optional `UsageMultiplier` for electric vehicles described using `Vehicles`.
 - Improves water heater tank losses when using `EnergyFactor` as the metric; now consistent with how `UniformEnergyFactor` is handled.
 
 __Bugfixes__